Disable the use-xdg-base-directories option
This setting theoretically should make Nix to conform to XDG
specification when storing files in the home directory.

The following paths should be affected (assuming default XDG paths):
- `~/.nix-profile` -> `~/.local/state/nix/profile`
- `~/.nix-defexpr` -> `~/.local/state/nix/defexpr`
- `~/.nix-channels` -> `~/.local/state/nix/channels`

Unfortunatelly, the support for this option in nix-darwin and
Home Manager is lacking. This is especially true when using a
standalone Home Manager setup.

In nix-darwin the XDG profile directory is not included in the
`environment.profiles` option, which results in the PATH variable
being set incorrectly.

In a standalone Home Manager setup the `use-xdg-base-directories`
option must also be enabled, even if we don't want HM to manage Nix
configuration. This is required, so the `home.profileDirectory` option
is set correctly to `~/.local/state/nix/profile`. On the other hand,
setting this option means that HM will create unnecessary configuration
file in `~/.config/nix/nix.conf`. Additionally, annoying warrnings are
emmited every time the `switch` command is run (because the non-root
user is not allowed to set this option).

Also, some tools rely on the old locations, because paths are hardcoded
in code.

Overall, it's a big mess and for now it's better to disable this option.
In practice, when this option is turned off, Nix still stores files in
`~/.local/state/nix`, but it creates symlinks in the home directory.

Some related issues for a reference:
https://github.com/LnL7/nix-darwin/issues/943
https://github.com/LnL7/nix-darwin/issues/947
https://github.com/nix-community/home-manager/issues/4807
https://github.com/nix-community/home-manager/issues/5561
https://github.com/nix-community/home-manager/issues/4593
https://github.com/nix-community/home-manager/issues/5805