Add session management for Neovim
Both mini.sessions and persistence.nvim are configured.
Currently, mini.session is disabled in favour of persistence.nvim.

The mini.session plugin allows for more control and flexibility, but it
has limited ability to restore session for cwd. It always restores last
global session, unless local session is explicitly created. The problem
with local sessions is that only one local session is allowed for a
given directory. Also, local sesssions are always stored in a project
directory.